% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_url.R
\docType{data}
\name{default_url}
\alias{default_url}
\title{Base NHSE statistics url}
\format{
An object of class \code{character} of length 1.
}
\usage{
default_url
}
\description{
Expected to be used unless otherwise specified, this allows the config
to just specify a path relative to the work areas page.
}
\keyword{internal}
