% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links_by_regex.R
\name{links_by_regex}
\alias{links_by_regex}
\title{Nested search for links by regex pattern}
\usage{
links_by_regex(base_url, patterns)
}
\arguments{
\item{base_url}{(chr) String url to search for patterns.}

\item{patterns}{(chr) Vector of regex patterns to search in nested pattern.}
}
\value{
(chr) Vector of filtered link strings.
}
\description{
Nested search for links by regex pattern
}
\keyword{internal}
