% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links_in_page.R
\name{links_in_page}
\alias{links_in_page}
\title{Extract all links from a web page}
\usage{
links_in_page(url)
}
\arguments{
\item{url}{Web address to search for links}
}
\value{
(chr) Vector of link strings
}
\description{
Parses a page using \link{rvest}, pulling the href attribute value from all \if{html}{\out{<a>}}
elements in the page. Only keeps strings starting with "https" so ignores
e.g. header links.
}
\keyword{internal}
