% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yaml_config.R
\name{load_yaml_config}
\alias{load_yaml_config}
\title{Load YAML file config}
\usage{
load_yaml_config(yaml_name)
}
\arguments{
\item{yaml_name}{(chr) String file name}
}
\value{
(list) Contents of the config file
}
\description{
Load YAML file config
}
\keyword{internal}
